move libvirt errors under virt-handler/errors
The errors can live in their own package to avoid circular
dependencies inside (and even outside of) virtwrap.
